Fix history order user
